fix(financial): preserve Decimal precision in calculations
Keep monetary ratios and exchange rates as Decimal throughout financial
calculations and persistence to avoid binary floating-point rounding.

- Compute CAGR fractional powers without Double conversion.
- Decode, cache, convert, and validate exchange rates as Decimal.
- Update dependent APIs, view models, tests, and documentation.

+++ AssetFlow/Services/BackupService+SupplementalParsing.swift
@@ -158,7 +158,7 @@ extension BackupService {
         continue
       }
       do {
-        let rates = try JSONDecoder().decode([String: Double].self, from: ratesData)
+        let rates = try JSONDecoder().decode([String: Decimal].self, from: ratesData)
         let hasInvalidRate = rates.contains {
           $0.key.isEmpty || !$0.value.isFinite || $0.value <= 0
         }
